// src/models/fleet/details-tabs-ids.ts
export enum DETAILS_TABS_IDS {
    MAIN_NOTES = 'main_notes',
    TEAM_NOTES = 'team_notes',
    DRIVER_NOTES = 'driver_notes',
    DOCUMENTS = 'documents',
    MAP = 'map',
    FINANCE = 'finance',
    LOADS = 'loads',
    MANIFESTS = 'manifests',
    WARRANTY = 'warranty',
    TRANSACTIONS_HISTORY = 'transactions_history',
    MEMBERS = 'members',
    NOTIFICATIONS = 'notifications'
}